chore: switch `zip` to `binstall-zip`
The previous version of the zip package has an outdated transitive Rust Crypto
dependency that conflicts with newer Rust Crytpo crates, so it's important to
update.

The master branch of the `zip` package already contains the new Rust Crytpo
dependencies, but it's been waiting for a release for three months now so this
commit switches to the updated `binstall-zip` fork by the `cargo-bins` team.

Fore more context, please see: https://github.com/roynalnaruto/svm-rs/pull/62